+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @workflow/ai
 
+## 4.0.1-beta.35
+
+### Patch Changes
+
+- Updated dependencies []:
+  - workflow@4.0.1-beta.35
+
+## 4.0.1-beta.34
+
+### Patch Changes
+
+- Updated dependencies []:
+  - workflow@4.0.1-beta.34
+
 ## 4.0.1-beta.33
 
 ### Patch Changes