+++ packages/astro/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @workflow/astro
 
+## 4.0.0-beta.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.1-beta.29
+  - @workflow/vite@4.0.0-beta.2
+
+## 4.0.0-beta.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/builders@4.0.1-beta.28
+  - @workflow/vite@4.0.0-beta.2
+
 ## 4.0.0-beta.10
 
 ### Patch Changes