+++ packages/web/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @workflow/web
 
+## 4.0.1-beta.19
+
+### Patch Changes
+
+- [#656](https://github.com/vercel/workflow/pull/656) [`ef22f82`](https://github.com/vercel/workflow/commit/ef22f82c9ead53744bac23fa12ed6bfbb1aba0bb) Thanks [@VaguelySerious](https://github.com/VaguelySerious)! - Allow resuming hooks with payloads from the UI
+
+## 4.0.1-beta.18
+
+### Patch Changes
+
+- [#646](https://github.com/vercel/workflow/pull/646) [`f396833`](https://github.com/vercel/workflow/commit/f39683370dc187273bd8aa5108e11e49dffe027a) Thanks [@adriandlam](https://github.com/adriandlam)! - Fix missing next.config.ts inside built @workflow/web package
+
 ## 4.0.1-beta.17
 
 ### Patch Changes