+++ packages/world-local/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @workflow/world-local
 
+## 4.0.1-beta.20
+
+### Patch Changes
+
+- [#662](https://github.com/vercel/workflow/pull/662) [`f2d5997`](https://github.com/vercel/workflow/commit/f2d5997b800d6c474bb93d4ddd82cf52489752da) Thanks [@VaguelySerious](https://github.com/VaguelySerious)! - Throw an error when trying writing JSON that fails entity validation, and remove error when trying to read JSON that fails validation, replacing it with a warning. This unblocks UI/CLI when data is invalid.
+
 ## 4.0.1-beta.19
 
 ### Patch Changes